Moving files to github from local version control
Most of these files are in the state I left them in 2008, as I was porting
the DYCAST system from SmallWorld Magik to Python and PostGIS. A few
changes I have made more recently. The bulk of the core functionality of
the system is now present, although there are many features still missing.